refactor: frontend/src/socket 파일 frontend/src/business/services로 이동
- 새로 결정한 폴더 구조에 맞게 socket 파일을 이동해주었고, @business path alias도 설정해주었다.
