+++ frontend/tsconfig.json
@@ -25,7 +25,7 @@
       "@/*": ["*"],
       "@pages/*": ["pages/*"],
       "@components/*": ["components/*"],
-      "@assets/*": ["assets/*"],
+      "@business/*": ["business/*"]
     }
   },
   "include": ["src", "vite-env.d.ts"],
